Fixed bug in test_sge_test whereby it tried to write to Tom's home directory :-) Added progressive timeouts - if the cluster is happy the test will be faster. If it is slow, it wauits longer and eventually prints a qstat.

+++ tests/test_sge_jobs.py
@@ -23,12 +23,24 @@ def test_valid_command(dag_id, job_list_manager_sge):
         "sge_foobar", slots=2, mem_free=4, project="proj_qlogins",
         max_attempts=3)
     job_list_manager_sge.queue_job(job_id)
-    sleep(60)  # Give some time for the job to get to the executor
-    done = job_list_manager_sge.get_new_done()
-    fail_msg = ("Test failed: check that you have permission to run under "
-                "'proj_qlogins' and that there are available jobs under this"
-                " project")
-    assert len(done) == 1, fail_msg
+
+    total_sleep = 0
+    step_size = 10
+    while True:
+        sleep(step_size)  # Give some time for the job to get to the executor
+        total_sleep += step_size
+        done = job_list_manager_sge.get_new_done()
+        if len(done) == 1:
+            # Success
+            assert True
+            return
+        elif total_sleep > 180:
+            print(sge.qstat())
+            fail_msg = ("Test failed due to 180s timeout: check the previous qtsat output; "
+                        "check that you have permission to run under "
+                        "'proj_qlogins' and that there are available jobs under this"
+                        " project")
+            assert len(done) == 1, fail_msg
 
 
 def test_context_args(jsm_jqs, job_list_manager_sge):
@@ -38,17 +50,31 @@ def test_context_args(jsm_jqs, job_list_manager_sge):
         "sge_foobar", slots=2, mem_free=4, max_attempts=3,
         context_args={'sge_add_args': '-a {}'.format(delay_to)})
     job_list_manager_sge.queue_job(job_id)
-    sleep(60)  # Give some time for the job to get instantiated
-    with session_scope() as session:
-        jis = session.query(JobInstance).filter_by(job_id=job_id).all()
-        njis = len(jis)
-        status = jis[0].status
-        sge_jid = jis[0].executor_id
 
-    # Make sure the job actually got to SGE
-    assert njis == 1
-    # Make sure it hasn't advanced to running (i.e. the -a argument worked)
-    assert status == JobInstanceStatus.SUBMITTED_TO_BATCH_EXECUTOR
+    total_sleep = 0
+    step_size = 10
+    while True:
+        sleep(step_size)  # Give some time for the job to get to the executor
+        total_sleep += step_size
+        with session_scope() as session:
+            jis = session.query(JobInstance).filter_by(job_id=job_id).all()
+            njis = len(jis)
+            status = jis[0].status
+            sge_jid = jis[0].executor_id
+        # Make sure the job actually got to SGE
+        if njis == 1:
+            # Make sure it hasn't advanced to running (i.e. the -a argument worked)
+            assert status == JobInstanceStatus.SUBMITTED_TO_BATCH_EXECUTOR
+            # Cleanup
+            sge.qdel(sge_jid)
+            return
+        elif total_sleep > 180:
+            # done waiting
+            print(sge.qstat())
+            fail_msg = ("Test failed due to 180s timeout: check the  qtsat output; "
+                        "check that you have permission to run under "
+                        "'proj_qlogins' and that there are available jobs under this"
+                        " project")
+            assert njis == 1, fail_msg
 
-    # Cleanup
-    sge.qdel(sge_jid)
+
